API: Use square function for MTF calc instead of sine
Because:
  - The Siemens Star that we are fitting is a square function
  - This prevents MTF values larger than 1

Swapped out the sin + cos function for computing the MTF by fitting a
function to the image of the Siemens star. Replaced it with a square
function from scipy.signal because this is the actual expected signal
from the image. The sine functions would cause MTF values higher than
unity because they have to be curved.